refactor: health routes
